+++ tests/unit/test_external_bangs.py
@@ -1,6 +1,6 @@
 from searx.external_bang import get_node, resolve_bang_definition, get_bang_url, get_bang_definition_and_autocomplete
 from searx.search import SearchQuery, EngineRef
-from searx.testing import SearxTestCase
+from tests import SearxTestCase
 
 
 TEST_DB = {
